Meta-function

Meta-function

electrify(app_name = NULL, product_name = "product_name",
  short_description = NULL, semantic_version = NULL,
  build_path = NULL, mran_date = NULL, cran_like_url = NULL,
  function_name = NULL, git_host = NULL, git_repo = NULL,
  local_package_path = NULL, package_install_opts = NULL,
  run_build = TRUE, nodejs_path = file.path(system.file(package =
  "electricShine"), "nodejs"), nodejs_version = "v10.16.0",
  permission = FALSE)

Arguments

app_name
This will be the name of the executable. It's a uniform type identifier (UTI) that contains only alphanumeric (A-Z,a-z,0-9), hyphen (-), and period (.) characters. see https://www.electron.build/configuration/configuration and https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html#//apple_ref/doc/uid/20001431-102070

product_name
String - allows you to specify a product name for your executable which contains spaces and other special characters not allowed in the name property. https://www.electron.build/configuration/configuration

short_description
short app description

semantic_version
semantic version of your app, as character (not numeric!); See https://semver.org/ for more info on semantic versioning.

build_path
Path where th bbuild files will be created, preferably points to an empty directory. Must not contain a folder with the name as what you put for electrify(app_name).

mran_date
MRAN snapshot date, formatted as 'YYYY-MM-DD'

cran_like_url
url to cran-like repository

function_name
the function name in your package that starts the shiny app

git_host
one of c("github", "gitlab", "bitbucket")

git_repo
GitHub/Bitbucket/GitLab username/repo of your the shiny-app package (e.g. 'chasemc/demoAPP'). Can also use notation for commits/branch (i.e. "chasemc/demoapp@d81fff0).

local_package_path
path to local shiny-app package, if 'git_package' isn't used

package_install_opts
optional arguments passed to remotes::install_github, install_gitlab, install_bitbucket, or install_local

run_build
logical, whether to start the build process, helpful if you want to modify anthying before building

nodejs_path
path to nodejs

nodejs_version
nodejs version to install

permission
automatically grant permission to install nodejs and R