Modify background.js to include the call to the shiny app

Modify background.js to include the call to the shiny app

modify_background_js(background_js_path, my_package_name, function_name,
  r_path)

Arguments

background_js_path
path to the final background.js, not the one in inst/...

my_package_name
package name, will be used for namespacing- (e.g. 'dplyr' in 'dplyr::filter()')

function_name
function that runs your shiny app - (e.g. 'filter' in 'dplyr::filter()')

r_path
path from "r_lang" folder to the R/Rscript executable

Value

none, side effect